requirements_multiview.txt
fal-client
pyyaml
requests
